iniciando cupom de produto

+++ src/main/resources/db/migration/V0004__create_db_cupom.sql
@@ -1,7 +1,7 @@
 CREATE TABLE cupom (
 	cupom_id INTEGER NOT NULL,
 	produto_id INTEGER NOT NULL,
-	quant_vendida DECIMAL(9,2) NOT NULL,
+	quant_vendida INTEGER NOT NULL,
 	valor_produto DECIMAL(9,2) NOT NULL,
 	PRIMARY KEY (cupom_id, produto_id)
 ) engine=InnoDB DEFAULT CHARSET=utf8;
